spec/requirements.txt
yamale>=1.5
PyYAML>=3.12
